refactor: Dealer와 Player의 부모 객체 생성
