+++ src/main/java/domain/player/Dealer.java
@@ -1,13 +1,13 @@
 package domain.player;
 
-public class Dealer extends Player {
+public class Dealer extends Participant {
 
     public Dealer(String name) {
         super(name);
     }
 
     public boolean isTotalValue16OrLess() {
-        return getTotalValue() <= 16;
+        return getTotalScore() <= 16;
     }
 
     public String getFirstCardStatus() {
