+++ src/main/java/domain/player/attribute/Score.java
@@ -1,12 +1,12 @@
-package domain.player;
+package domain.player.attribute;
 
 import domain.card.Card;
+import static util.Constants.ACE_SCORE;
+import static util.Constants.DEALER_REFERENCE_POINT;
 
 public class Score {
 
     private int score = 0;
-
-    private static final int ACE_SCORE = 11;
 
     public void addValue(Card card){
         if (card.isAce()) {
@@ -16,16 +16,15 @@ public class Score {
         score += card.getValue();
     }
 
-    public int getScore() {
-        return score;
-    }
-
     private void calculateAceCard(Card card) {
-        if (score + ACE_SCORE > 21) {
+        if (score + ACE_SCORE > DEALER_REFERENCE_POINT) {
             score += card.getValue();
             return;
         }
         score += ACE_SCORE;
     }
 
+    public int getScore() {
+        return score;
+    }
 }
