+++ src/test/java/domain/player/CardStatusTest.java
@@ -1,6 +1,7 @@
 package domain.player;
 
 import domain.card.Card;
+import domain.player.attribute.CardStatus;
 import java.util.Arrays;
 import java.util.List;
 import org.assertj.core.api.Assertions;
