+++ src/test/java/domain/player/NameTest.java
@@ -1,5 +1,6 @@
 package domain.player;
 
+import domain.player.attribute.Name;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
